+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Yourloops data visualization for diabetes device data
 This node package is a dependency of Blip and was forked from tidepool/viz.
 
+## 0.6.0 - 2019-10-30
+### Added
+- [PT-131] Make Device Settings i18n compatible so that they will be displayed in blip in user language.
+
 ## 0.5.0 - 2019-10-14
 ### Added
 - [PT-573] Integrate latest changes from Tidepool Viz [v1.8.0](https://github.com/tidepool-org/viz/releases/tag/v1.8.0)